Fix AudioListController@destroy

+++ app/Http/Controllers/AudioListController.php
@@ -80,14 +80,20 @@ class AudioListController extends Controller
     public function destroy(Request $request, $lang, $swiss_number, $audio_id)
     {
         $audio = Audio::find($audio_id);
-        $dir_path = '/home/louis/taxitakemeto-core/public';
+        if ($audio != NULL) {
+            $dir_path = '/home/louis/taxitakemeto-core/public';
 
-        if (AudioListEditFacet::find($swiss_number) != NULL) {
-            if (file_exists($dir_path . $audio->path)) {
-                unlink($dir_path . $request->audio_path);
-                $audio->delete();
+            if (AudioListEditFacet::find($swiss_number) != NULL) {
+                if (file_exists($dir_path . $audio->path)) {
+                    unlink($dir_path . $request->audio_path);
+                    $audio->delete();
+                    return redirect("/$lang/audiolist_edit/$swiss_number", 303);
+                } else {
+                    return response(view('404'), 404);
+                }
+            } else {
+                return response(view('404'), 404);
             }
-            return redirect("/$lang/audiolist_edit/$swiss_number", 303);
         } else {
             return response(view('404'), 404);
         }
